[kotlin] Introduce ConvertToBlockBodyUtils
