+++ plugins/kotlin/idea/src/org/jetbrains/kotlin/idea/intentions/ConvertToBlockBodyIntention.kt
@@ -3,23 +3,19 @@
 package org.jetbrains.kotlin.idea.intentions
 
 import com.intellij.openapi.editor.Editor
-import com.intellij.psi.PsiComment
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiWhiteSpace
 import org.jetbrains.kotlin.descriptors.CallableDescriptor
 import org.jetbrains.kotlin.idea.base.resources.KotlinBundle
 import org.jetbrains.kotlin.idea.caches.resolve.analyze
 import org.jetbrains.kotlin.idea.caches.resolve.resolveToDescriptorIfAny
 import org.jetbrains.kotlin.idea.codeinsight.api.classic.intentions.SelfTargetingIntention
-import org.jetbrains.kotlin.idea.codeinsight.utils.adjustLineIndent
-import org.jetbrains.kotlin.idea.core.setType
+import org.jetbrains.kotlin.idea.codeinsight.utils.ConvertToBlockBodyContext
+import org.jetbrains.kotlin.idea.codeinsight.utils.ConvertToBlockBodyUtils
+import org.jetbrains.kotlin.idea.core.ShortenReferences
 import org.jetbrains.kotlin.idea.intentions.ConvertToBlockBodyIntention.Holder.convert
-import org.jetbrains.kotlin.idea.intentions.ConvertToBlockBodyIntention.Holder.returnType
-import org.jetbrains.kotlin.idea.util.resultingWhens
+import org.jetbrains.kotlin.idea.intentions.ConvertToBlockBodyIntention.Holder.createContext
+import org.jetbrains.kotlin.idea.util.IdeDescriptorRenderers
 import org.jetbrains.kotlin.psi.*
-import org.jetbrains.kotlin.psi.psiUtil.endOffset
-import org.jetbrains.kotlin.psi.psiUtil.siblings
-import org.jetbrains.kotlin.psi.psiUtil.startOffset
 import org.jetbrains.kotlin.types.KotlinType
 import org.jetbrains.kotlin.types.isError
 import org.jetbrains.kotlin.types.isNullabilityFlexible
@@ -31,21 +27,7 @@ class ConvertToBlockBodyIntention : SelfTargetingIntention<KtDeclarationWithBody
     KtDeclarationWithBody::class.java,
     KotlinBundle.lazyMessage("convert.to.block.body")
 ) {
-    override fun isApplicableTo(element: KtDeclarationWithBody, caretOffset: Int): Boolean {
-        if (element is KtFunctionLiteral || element.hasBlockBody() || !element.hasBody()) return false
-
-        when (element) {
-            is KtNamedFunction -> {
-                val returnType = element.returnType() ?: return false
-                if (!element.hasDeclaredReturnType() && returnType.isError) return false// do not convert when type is implicit and unknown
-                return true
-            }
-
-            is KtPropertyAccessor -> return true
-
-            else -> error("Unknown declaration type: $element")
-        }
-    }
+    override fun isApplicableTo(element: KtDeclarationWithBody, caretOffset: Int) = createContext(element) != null
 
     override fun skipProcessingFurtherElementsAfter(element: PsiElement) =
         element is KtDeclaration || super.skipProcessingFurtherElementsAfter(element)
@@ -56,85 +38,47 @@ class ConvertToBlockBodyIntention : SelfTargetingIntention<KtDeclarationWithBody
 
     object Holder {
         fun convert(declaration: KtDeclarationWithBody, withReformat: Boolean = false): KtDeclarationWithBody {
-            val body = declaration.bodyExpression!!
-            val prevComments = body.comments(next = false)
-            val nextComments = body.comments(next = true)
-
-            fun generateBody(returnsValue: Boolean): KtExpression {
-                val bodyType = body.analyze().getType(body)
-                val psiFactory = KtPsiFactory(declaration.project)
-                if (bodyType != null && bodyType.isUnit() && body is KtNameReferenceExpression) return psiFactory.createEmptyBody()
-                val unitWhenAsResult = (bodyType == null || bodyType.isUnit()) && body.resultingWhens().isNotEmpty()
-                val needReturn = returnsValue && (bodyType == null || (!bodyType.isUnit() && !bodyType.isNothing()))
-                val newBody = if (needReturn || unitWhenAsResult) {
-                    val annotatedExpr = body as? KtAnnotatedExpression
-                    val returnedExpr = annotatedExpr?.baseExpression ?: body
-                    val block = psiFactory.createSingleStatementBlock(psiFactory.createExpressionByPattern("return $0", returnedExpr))
-                    val statement = block.firstStatement
-                    annotatedExpr?.annotationEntries?.forEach {
-                        block.addBefore(it, statement)
-                        block.addBefore(psiFactory.createNewLine(), statement)
-                    }
-                    block
-                } else {
-                    psiFactory.createSingleStatementBlock(body)
-                }
-                prevComments
-                    .dropWhile { it is PsiWhiteSpace }
-                    .forEach { newBody.addAfter(it, newBody.lBrace) }
-                nextComments
-                    .reversed()
-                    .forEach { newBody.addAfter(it, newBody.firstStatement) }
-                return newBody
-            }
-
-            val newBody = when (declaration) {
-                is KtNamedFunction -> {
-                    val returnType = declaration.returnType()!!
-                    if (!declaration.hasDeclaredReturnType() && !returnType.isUnit()) {
-                        declaration.setType(returnType)
-                    }
-                    generateBody(!returnType.isUnit() && !returnType.isNothing())
-                }
-
-                is KtPropertyAccessor -> {
-                    val parent = declaration.parent
-                    if (parent is KtProperty && parent.typeReference == null) {
-                        val descriptor = parent.resolveToDescriptorIfAny()
-                        (descriptor as? CallableDescriptor)?.returnType?.let { parent.setType(it) }
-                    }
-
-                    generateBody(declaration.isGetter)
-                }
-
-                else -> throw RuntimeException("Unknown declaration type: $declaration")
-            }
-
-            declaration.equalsToken!!.delete()
-            prevComments.filterIsInstance<PsiComment>().forEach { it.delete() }
-            nextComments.forEach { it.delete() }
-            val replaced = body.replace(newBody)
-            if (withReformat) {
-                declaration.containingKtFile.adjustLineIndent(replaced.startOffset, replaced.endOffset)
-            }
+            val context = createContext(declaration, withReformat) ?: return declaration
+            ConvertToBlockBodyUtils.convert(declaration, context)
             return declaration
         }
 
-        internal fun KtNamedFunction.returnType(): KotlinType? {
-            val descriptor = resolveToDescriptorIfAny()
-            val returnType = descriptor?.returnType ?: return null
-            if (returnType.isNullabilityFlexible()
-                && descriptor.overriddenDescriptors.firstOrNull()?.returnType?.isMarkedNullable == false
-            ) return returnType.makeNotNullable()
-            return returnType
+        fun createContext(declaration: KtDeclarationWithBody, reformat: Boolean = false): ConvertToBlockBodyContext? {
+            if (!ConvertToBlockBodyUtils.isConvertibleByPsi(declaration)) return null
+
+            val body = declaration.bodyExpression ?: return null
+
+            val returnType = declaration.returnType() ?: return null
+            if (returnType.isError && declaration is KtNamedFunction && !declaration.hasDeclaredReturnType()) return null
+
+            val bodyType = body.analyze().getType(body)
+
+            return ConvertToBlockBodyContext(
+                returnTypeIsUnit = returnType.isUnit(),
+                returnTypeIsNothing = returnType.isNothing(),
+                returnTypeString = IdeDescriptorRenderers.SOURCE_CODE.renderType(returnType),
+                bodyTypeIsUnit = bodyType?.isUnit() == true,
+                bodyTypeIsNothing = bodyType?.isNothing() == true,
+                reformat = reformat,
+                shortenReferences = { ShortenReferences.DEFAULT.process(it) }
+            )
         }
 
-        private fun KtExpression.comments(next: Boolean): List<PsiElement> = siblings(forward = next, withItself = false)
-            .takeWhile { it is PsiWhiteSpace || it is PsiComment }
-            .takeIf { it.hasComment() }
-            .orEmpty()
-            .toList()
-
-        private fun Sequence<PsiElement>.hasComment(): Boolean = any { it is PsiComment }
+        private fun KtDeclarationWithBody.returnType(): KotlinType? {
+            return when (this) {
+                is KtNamedFunction -> {
+                    val descriptor = resolveToDescriptorIfAny()
+                    val returnType = descriptor?.returnType ?: return null
+                    if (returnType.isNullabilityFlexible()
+                        && descriptor.overriddenDescriptors.firstOrNull()?.returnType?.isMarkedNullable == false
+                    ) returnType.makeNotNullable() else returnType
+                }
+                is KtPropertyAccessor -> {
+                    val parent = this.parent as? KtProperty
+                    (parent?.resolveToDescriptorIfAny() as? CallableDescriptor)?.returnType
+                }
+                else -> null
+            }
+        }
     }
 }
