--- plugins/kotlin/idea/tests/testData/intentions/convertToBlockBody/ifWhenUnit.kt.after.k2
@@ -1,14 +1,0 @@
-enum class AccessMode { READ, WRITE, EXECUTE }
-
-fun foo() {}
-
-fun whenExpr(access: AccessMode, arg: Boolean) {
-    if (arg) {} else {
-        foo()
-        when (access) {
-            AccessMode.READ -> {}
-            AccessMode.WRITE -> {}
-            AccessMode.EXECUTE -> {}
-        }
-    }
-}
