+++ plugins/kotlin/idea/tests/testData/intentions/convertToBlockBody/whenUnit.kt.after
@@ -1,7 +1,7 @@
 enum class AccessMode { READ, WRITE, EXECUTE }
 
 fun whenExpr(access: AccessMode) {
-    return when (access) {
+    when (access) {
         AccessMode.READ -> {}
         AccessMode.WRITE -> {}
         AccessMode.EXECUTE -> {}
