--- plugins/kotlin/idea/tests/testData/intentions/convertToBlockBody/whenUnit.kt.after.k2
@@ -1,9 +1,0 @@
-enum class AccessMode { READ, WRITE, EXECUTE }
-
-fun whenExpr(access: AccessMode) {
-    when (access) {
-        AccessMode.READ -> {}
-        AccessMode.WRITE -> {}
-        AccessMode.EXECUTE -> {}
-    }
-}
